Move echo functions to utils.
