+++ hatch/utils.py
@@ -8,6 +8,7 @@ from datetime import datetime
 from contextlib import contextmanager
 from pathlib import Path
 from tempfile import TemporaryDirectory
+import click
 
 ON_WINDOWS = False
 if os.name == 'nt' or platform.system() == 'Windows':  # no cov
@@ -20,6 +21,35 @@ VENV_FLAGS = {
     'VIRTUAL_ENV',
     'CONDA_PREFIX'
 }
+
+CONTEXT_SETTINGS = {
+    'help_option_names': ['-h', '--help'],
+}
+
+UNKNOWN_OPTIONS = {
+    'ignore_unknown_options': True,
+    **CONTEXT_SETTINGS
+}
+
+
+def echo_success(text, nl=True):
+    click.secho(text, fg='cyan', bold=True, nl=nl)
+
+
+def echo_failure(text, nl=True):
+    click.secho(text, fg='red', bold=True, nl=nl)
+
+
+def echo_warning(text, nl=True):
+    click.secho(text, fg='yellow', bold=True, nl=nl)
+
+
+def echo_waiting(text, nl=True):
+    click.secho(text, fg='magenta', bold=True, nl=nl)
+
+
+def echo_info(text, nl=True):
+    click.secho(text, fg='white', bold=True, nl=nl)
 
 
 def venv_active():
